Move sidebar config to a separate class
