+++ config/application.rb
@@ -46,6 +46,7 @@ module Publify
 
   # Load included libraries.
   require 'sidebar'
+  require 'sidebar_configuration'
   require 'sidebar_registry'
   require 'publify_sidebar'
   require 'publify_textfilters'
