+++ lib/archives_sidebar/lib/archives_sidebar.rb
@@ -1,4 +1,4 @@
-class ArchivesSidebar < Sidebar
+class ArchivesSidebar < SidebarConfiguration
   description 'Displays links to monthly archives'
   setting :title, 'Archives'
   setting :show_count, true, label: 'Show article counts', input_type: :checkbox
