+++ lib/generators/sidebar/templates/sidebar.rb.erb
@@ -1,4 +1,4 @@
-class <%= class_name %> < Sidebar
+class <%= class_name %> < SidebarConfiguration
   # display_name "<%= class_name.underscore.humanize %>" # Default
   description "Describe your sidebar here"
