+++ lib/livesearch_sidebar/lib/livesearch_sidebar.rb
@@ -1,4 +1,4 @@
-class LivesearchSidebar < Sidebar
+class LivesearchSidebar < SidebarConfiguration
   description 'Adds livesearch to your Publify blog'
 
   setting :title, 'Search'
