+++ lib/meta_sidebar/lib/meta_sidebar.rb
@@ -1,5 +1,4 @@
-# coding: utf-8
-class MetaSidebar < Sidebar
+class MetaSidebar < SidebarConfiguration
   description "This widget just displays links to Publify main site, this blog's admin and RSS."
 
   setting :title, 'Meta'
