+++ lib/notes_sidebar/lib/notes_sidebar.rb
@@ -1,4 +1,4 @@
-class NotesSidebar < Sidebar
+class NotesSidebar < SidebarConfiguration
   description 'Displays the latest notes'
   setting :title, 'Notes'
   setting :count, 5, label: 'Number of notes'
