+++ lib/page_sidebar/lib/page_sidebar.rb
@@ -1,4 +1,4 @@
-class PageSidebar < Sidebar
+class PageSidebar < SidebarConfiguration
   display_name 'Page'
   description 'Show pages for this blog'
