+++ lib/popular_sidebar/lib/popular_sidebar.rb
@@ -1,4 +1,4 @@
-class PopularSidebar < Sidebar
+class PopularSidebar < SidebarConfiguration
   description 'Displays the most popular posts'
   setting :title, 'Most popular'
   setting :count, 5, label: 'Number articles'
