+++ lib/tag_sidebar/lib/tag_sidebar.rb
@@ -1,4 +1,4 @@
-class TagSidebar < Sidebar
+class TagSidebar < SidebarConfiguration
   display_name 'Tags'
   description 'Show most popular tags for this blog'
