+++ lib/xml_sidebar/lib/xml_sidebar.rb
@@ -1,4 +1,4 @@
-class XmlSidebar < Sidebar
+class XmlSidebar < SidebarConfiguration
   display_name 'XML Syndication'
   description 'RSS and Atom feeds'
